Added import and export tasks for individual models Fixed error when setting @@xmlpath and @@models Changed default xmlpath

+++ lib/ez_import/main.rb
@@ -1,15 +1,17 @@
-module EzImport
 
+module EzImport
+  require 'hpricot'
   class Main
 
     @@models = []
-    @@xmlpath = "db/ezimport"
+    @@xmlpath = "db/ez_import"
 
     cattr_accessor :models
     cattr_accessor :xmlpath
 
-    def self.export(model_name)
-      model_name, model_obj = self.get_model(model_name)
+    def self.export(model_name_original)
+      model_name, model_obj = self.get_model(model_name_original)
+      raise "Model '#{model_name_original}' not found" if model.nil?
       unless File.directory?(@@xmlpath)
         Dir.mkdir(@@xmlpath)
         puts "create #{@@xmlpath}"
@@ -20,19 +22,24 @@ module EzImport
     end
 
     def self.get_model(model_name)
-      Dir.glob("#{RAILS_ROOT}/app/models/**/*rb").each{|m| require_or_load m }
+      model_name = model_name.downcase.gsub('_','')
+      Dir.glob("#{Rails.root}/app/models/**/*rb").each{|m| require_or_load m }
       model_list = {}
       ActiveRecord::Base.subclasses.each {|m| model_list[m.to_s.downcase] = m}
-      [model_list[model_name.downcase].to_s.underscore, model_list[model_name.downcase]]
+      m1 = model_list[model_name].to_s.underscore
+      m2 = model_list[model_name]
+      [m1, m2]
     end
 
-    def self.import(model_name)
-      model_name, model = self.get_model(model_name)
+    def self.import(model_name_original)
+      model_name, model = self.get_model(model_name_original)
+      raise "Model '#{model_name_original}' not found" if model.nil?
       model.delete_all
-      puts "\n\n"
+      puts "\n"
       puts "Load #{model_name.pluralize}.xml"
       file_content = File.read("#{@@xmlpath}/#{model_name.pluralize}.xml")
       file_content.gsub!("&amp;", "&")
+      count = 0
       Hpricot(file_content).search(model_name.gsub('_', '-')).each do |record|
         model.new do |new_instance|
           model.columns.each do |col|
@@ -41,8 +48,9 @@ module EzImport
           end
           new_instance.save
         end
-        puts "Added #{model_name}: #{(record/:name).innerHTML}" unless (record/:name).innerHTML.blank?
+        count += 1
       end
+      puts "\t#{count} records loaded"
       if model.methods.include?(:after_ezimport)
         puts "Running #{model_name}.after_ezimport..."
         model.after_ezimport
